requirements.txt
#Use pip3 -r requirements.txt to install the dependencies
Pillow>=5.4.1
drawille>=0.1.0
npyscreen>=4.10.5
